LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/sandboxed-iframe-navigator-serviceworker.https-expected.txt


PASS Accessing navigator.serviceWorker in normal iframe should not throw. 
FAIL Accessing navigator.serviceWorker in sandboxed iframe should throw. assert_equals: expected "navigator.serviceWorker failed: SecurityError" but got "ok"
PASS Accessing navigator.serviceWorker in sandboxed iframe with allow-same-origin flag should not throw. 
FAIL Switching iframe sandbox attribute while loading the iframe assert_equals: expected "navigator.serviceWorker failed: SecurityError" but got "ok"
